More Python Script work
The script now outputs to a csv with all game data from the previous season. This can only be modified by changing the hard coded URL. I will fix that in the future. Now it is time to move onto the R side of the project and see what we can figure out about this data.